Removed broken Issuer abstraction from core library.
The Issuer class failed to take into account the possibility of several 
default-probability curves and recovery rates for each issuer, based on
seniority or currency.
In most instrument and pricing engines, instances of Issuer were replaced
by the data members it contained.  The Issuer class itself was moved to
ql/experimental/credit to support the SyntheticCDO class, which uses it
for the Pool and Basket classes and might used as a case study for finding
the correct abstraction.



git-svn-id: https://quantlib.svn.sourceforge.net/svnroot/quantlib/trunk/QuantLib@15962 8618b1d8-e22c-0410-b026-96a5dba3e089