Added support for VC++2010 to the core library.
The support is still incomplete. On the one hand, not all warnings were removed.
On the other hand, VC10 doesn't compile parts of Boost::uBLAS 1.44.  For the time
being, QL_NO_UBLAS_SUPPORT was defined in config.msvc.hpp (which disables a few
methods and classes.)  The define should be removed once everything works.

git-svn-id: https://quantlib.svn.sourceforge.net/svnroot/quantlib/trunk/QuantLib@17428 8618b1d8-e22c-0410-b026-96a5dba3e089