Moved forecastFixing(Date,Date,Time) overload to private section.
When used with cached dates, this overload can give quite a
performance boost to coupon calculations (see IborCoupon)
but is potentially misleading: by passing the wrong dates,
one can ask a 6-months index for a 1-year fixing.

For that reason, we're making this method private and
we're declaring the IborCoupon class (which uses it) as a
friend.

git-svn-id: https://quantlib.svn.sourceforge.net/svnroot/quantlib/trunk/QuantLib@18058 8618b1d8-e22c-0410-b026-96a5dba3e089