Better detection of forecast/past fixings for inflation indexes.
When an interpolated index is asked for a fixing at the beginning
of a month, the fixing for the following (which would have zero
weight in the interpolation) is no longer required.
Also, if a fixing is loaded in the index time series, it can be
used even its observation lag has not fully elapsed.

git-svn-id: https://quantlib.svn.sourceforge.net/svnroot/quantlib/trunk/QuantLib@18065 8618b1d8-e22c-0410-b026-96a5dba3e089