Relaxed extrapolation check for default-probability curves.
When calculating default probabilities between two dates or times,
allow the first to precede the reference date. This effectively
assumes that the default probability before the reference is null,
and helps in cases where a coupon protection extends a couple of
days before the reference due to adjustments (for instance, when
the protection starts on a Saturday and the reference is rolled to
the following Monday).

git-svn-id: https://quantlib.svn.sourceforge.net/svnroot/quantlib/trunk/QuantLib@18443 8618b1d8-e22c-0410-b026-96a5dba3e089