intraday patch
This set of patches implements support for intraday timestamps, and
creates a continuoustime day counter which allows for intraday
calculations.

There is one backward compatibility issue.  Although the new DCC
allows support for time computations in which the base unit is not
one year, the time computation is done with the method yearFraction
for backwards compatibility.  A new interface timeFraction has been
created to provide a better name for time computations.